(WIP) Fixing 6 car+ tables

+++ src/common/online.c
@@ -122,7 +122,8 @@ void reduce_digest(char *charDigest, unsigned int index, char *charPlain, int pw
     char_to_password("abcdefg", plainText, pwdLength);
 
     unsigned long long temp = 0;
-    temp = (unsigned long long) ((*digest).i[0] + (*digest).i[1] + (*digest).i[2] + (*digest).i[3] + index) %
+    temp = (unsigned long long) ((unsigned long long)(*digest).i[0] + (unsigned long long)(*digest).i[1] +
+            (unsigned long long)(*digest).i[2] + (unsigned long long)(*digest).i[3] + (unsigned long long)index) %
            (unsigned long long) (power(CHARSET_LENGTH, pwdLength));
 
     for (int i = pwdLength - 1; i >= 0; i--) {
